ai-service/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
transformers
pillow
numpy
python-multipart
